pyproject.toml
[project]
name = "linkedin"
version = "0.1.0"
description = "LinkedIn MCP server for Claude Desktop"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "linkedin-api>=2.3.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]
